feat: update analyzer to use golang.org/x/tools 0.33.0
This version is the one expected by the latest version of golangci-lint.
This is necessary to ensure compatibility of the rule with the linter.

The change occurred between golang.org/x/tools v0.24.0 and v0.25.0 with
the code behind https://pkg.go.dev/golang.org/x/tools/go/callgraph/static#CallGraph

Here is the commit that caused the issue

https://github.com/golang/tools/commit/c538e2c079ea0676702503ccfe54f8c3ddd7ae81

The refactoring was done to improve performance, but it also removed the
functionality we need for our analyzer.

We need to have the wrappers in the CallGraph to be able to check the
wrapped functions.

Because of that we had to copy the code that was present in
static.CallGraph before the refactoring.

https://github.com/golang/tools/blob/v0.24.0/go/callgraph/static/static.go#L19